feature/43-active-community-channel: move setActiveCommunity to CommunityService, move setActiveChannel to ChannelService
