+++ qiusuo-root/qiusuo-communityservice/src/main/java/com/qiusuo/communityservice/graphql/mutation/UserMutation.java
@@ -3,6 +3,8 @@ package com.qiusuo.communityservice.graphql.mutation;
 import com.qiusuo.communityservice.domain.model.Channel;
 import com.qiusuo.communityservice.domain.model.Community;
 import com.qiusuo.communityservice.domain.model.User;
+import com.qiusuo.communityservice.domain.service.ChannelService;
+import com.qiusuo.communityservice.domain.service.CommunityService;
 import com.qiusuo.communityservice.domain.service.UserService;
 import com.qiusuo.communityservice.graphql.types.UserRegistrationInput;
 import graphql.kickstart.tools.GraphQLMutationResolver;
@@ -11,9 +13,14 @@ import org.springframework.stereotype.Component;
 @Component
 public class UserMutation implements GraphQLMutationResolver {
     private UserService userService;
+    private ChannelService channelService;
+    private CommunityService communityService;
 
-    public UserMutation(UserService userService) {
+    public UserMutation(UserService userService, ChannelService channelService,
+                        CommunityService communityService) {
         this.userService = userService;
+        this.channelService = channelService;
+        this.communityService = communityService;
     }
 
     public User registerUser(UserRegistrationInput userRegistrationInput) {
@@ -21,10 +28,10 @@ public class UserMutation implements GraphQLMutationResolver {
     }
 
     public Community setActiveCommunity(String communityId) {
-        return userService.setActiveCommunity(communityId);
+        return communityService.setActiveCommunity(communityId);
     }
 
     public Channel setActiveChannel(String channelId) {
-        return userService.setActiveChannel(channelId);
+        return channelService.setActiveChannel(channelId);
     }
 }
